Qrank and Barista check that sp score, sp rank, b/y ions matched and b/y ions total in the search result file are provided. if those files are not provided, it doesn't print out them.| code reviewed by Sean.

+++ src/c/qranker-barista/CruxParser.cpp
@@ -68,7 +68,7 @@ void CruxParser ::readMatches(
       matches.sp_score.push_back(reader.getDouble(SP_SCORE_COL));
       matches.calc_mass.push_back(reader.getDouble(PEPTIDE_MASS_COL));
       matches.xcorr_score.push_back(reader.getDouble(XCORR_SCORE_COL));
-      matches.num_ions_matched.push_back(reader.getDouble(BY_IONS_MATCHED_COL));
+      matches.num_ions_matched.push_back(reader.getInteger(BY_IONS_MATCHED_COL));
       matches.num_total_ions.push_back(reader.getFloat(BY_IONS_TOTAL_COL)); 
         
       //set sequence_id
